refactor: DTO 클래스 분리
