+++ src/main/java/roomescape/member/controller/MemberController.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.global.auth.annotation.RoleRequired;
-import roomescape.member.dto.request.MemberRequest.MemberCreateRequest;
-import roomescape.member.dto.response.MemberResponse.MemberCreateResponse;
-import roomescape.member.dto.response.MemberResponse.MemberReadResponse;
+import roomescape.member.dto.request.MemberCreateRequest;
+import roomescape.member.dto.response.MemberCreateResponse;
+import roomescape.member.dto.response.MemberReadResponse;
 import roomescape.member.entity.RoleType;
 import roomescape.member.service.MemberService;
