--- src/main/java/roomescape/member/dto/request/AuthRequest.java
@@ -1,12 +1,0 @@
-package roomescape.member.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public class AuthRequest {
-
-    public record LoginRequest(
-            @NotNull String email,
-            @NotNull String password
-    ) {
-    }
-}
