+++ src/main/java/roomescape/member/dto/request/LoginRequest.java
@@ -1,0 +1,9 @@
+package roomescape.member.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record LoginRequest(
+        @NotNull String email,
+        @NotNull String password
+) {
+}
