+++ src/main/java/roomescape/member/dto/request/MemberCreateRequest.java
@@ -1,0 +1,15 @@
+package roomescape.member.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import roomescape.member.entity.Member;
+import roomescape.member.entity.RoleType;
+
+public record MemberCreateRequest(
+        @NotNull String name,
+        @NotNull String email,
+        @NotNull String password
+) {
+    public Member toEntity() {
+        return new Member(name, email, password, RoleType.USER);
+    }
+}
