--- src/main/java/roomescape/member/dto/response/AuthResponse.java
@@ -1,20 +1,0 @@
-package roomescape.member.dto.response;
-
-import roomescape.global.auth.dto.LoginMember;
-
-public class AuthResponse {
-
-    public record LoginResponse(
-            Long id,
-            String name,
-            String role
-    ) {
-        public static LoginResponse from(LoginMember loginMember) {
-            return new LoginResponse(
-                    loginMember.id(),
-                    loginMember.name(),
-                    loginMember.role().name()
-            );
-        }
-    }
-}
