+++ src/main/java/roomescape/member/service/AuthService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import roomescape.global.auth.dto.LoginMember;
 import roomescape.global.auth.util.JwtUtil;
 import roomescape.global.error.exception.BadRequestException;
-import roomescape.member.dto.request.AuthRequest.LoginRequest;
+import roomescape.member.dto.request.LoginRequest;
 import roomescape.member.entity.Member;
 import roomescape.member.repository.MemberRepository;
