+++ src/main/java/roomescape/reservation/dto/request/ReservationCreateRequest.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+import java.time.LocalDate;
+
+public record ReservationCreateRequest(
+        @NotNull LocalDate date,
+        @NotNull Long timeId,
+        @NotNull Long themeId
+) {
+}
