--- src/main/java/roomescape/reservation/dto/request/ReservationTimeRequest.java
@@ -1,16 +1,0 @@
-package roomescape.reservation.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-import java.time.LocalTime;
-import roomescape.reservation.entity.ReservationTime;
-
-public class ReservationTimeRequest {
-
-    public record ReservationTimeCreateRequest(
-            @NotNull LocalTime startAt
-    ) {
-        public ReservationTime toEntity() {
-            return new ReservationTime(startAt);
-        }
-    }
-}
