+++ src/main/java/roomescape/reservation/dto/response/AvailableReservationTimeResponse.java
@@ -1,0 +1,11 @@
+package roomescape.reservation.dto.response;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import java.time.LocalTime;
+
+public record AvailableReservationTimeResponse(
+        Long id,
+        @JsonFormat(pattern = "HH:mm") LocalTime startAt,
+        boolean alreadyBooked
+) {
+}
