+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -9,14 +9,14 @@ import roomescape.global.error.exception.ConflictException;
 import roomescape.global.error.exception.NotFoundException;
 import roomescape.member.entity.Member;
 import roomescape.member.repository.MemberRepository;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationAdminCreateRequest;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationCreateRequest;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationReadFilteredRequest;
-import roomescape.reservation.dto.response.ReservationResponse.ReservationAdminCreateResponse;
-import roomescape.reservation.dto.response.ReservationResponse.ReservationCreateResponse;
-import roomescape.reservation.dto.response.ReservationResponse.ReservationReadFilteredResponse;
-import roomescape.reservation.dto.response.ReservationResponse.ReservationReadMemberResponse;
-import roomescape.reservation.dto.response.ReservationResponse.ReservationReadResponse;
+import roomescape.reservation.dto.request.ReservationAdminCreateRequest;
+import roomescape.reservation.dto.request.ReservationCreateRequest;
+import roomescape.reservation.dto.request.ReservationReadFilteredRequest;
+import roomescape.reservation.dto.response.ReservationAdminCreateResponse;
+import roomescape.reservation.dto.response.ReservationCreateResponse;
+import roomescape.reservation.dto.response.ReservationReadFilteredResponse;
+import roomescape.reservation.dto.response.ReservationReadMemberResponse;
+import roomescape.reservation.dto.response.ReservationReadResponse;
 import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.entity.ReservationTime;
 import roomescape.reservation.repository.ReservationRepository;
