+++ src/main/java/roomescape/reservation/service/ReservationTimeService.java
@@ -6,10 +6,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import roomescape.global.error.exception.BadRequestException;
 import roomescape.global.error.exception.ConflictException;
-import roomescape.reservation.dto.request.ReservationTimeRequest.ReservationTimeCreateRequest;
-import roomescape.reservation.dto.response.ReservationTimeResponse.AvailableReservationTimeResponse;
-import roomescape.reservation.dto.response.ReservationTimeResponse.ReservationTimeCreateResponse;
-import roomescape.reservation.dto.response.ReservationTimeResponse.ReservationTimeReadResponse;
+import roomescape.reservation.dto.request.ReservationTimeCreateRequest;
+import roomescape.reservation.dto.response.AvailableReservationTimeResponse;
+import roomescape.reservation.dto.response.ReservationTimeCreateResponse;
+import roomescape.reservation.dto.response.ReservationTimeReadResponse;
 import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.entity.ReservationTime;
 import roomescape.reservation.repository.ReservationRepository;
