+++ src/test/java/roomescape/reservation/acceptance/ReservationTimeAcceptanceTest.java
@@ -16,7 +16,7 @@ import roomescape.helper.TestHelper;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
 import roomescape.member.repository.MemberRepository;
-import roomescape.reservation.dto.request.ReservationTimeRequest.ReservationTimeCreateRequest;
+import roomescape.reservation.dto.request.ReservationTimeCreateRequest;
 
 @SpringBootTest(webEnvironment = WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
