+++ src/test/java/roomescape/reservation/integration/ReservationIntegrationTest.java
@@ -17,9 +17,9 @@ import roomescape.global.error.exception.ConflictException;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
 import roomescape.member.repository.MemberRepository;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationAdminCreateRequest;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationCreateRequest;
-import roomescape.reservation.dto.request.ReservationRequest.ReservationReadFilteredRequest;
+import roomescape.reservation.dto.request.ReservationAdminCreateRequest;
+import roomescape.reservation.dto.request.ReservationCreateRequest;
+import roomescape.reservation.dto.request.ReservationReadFilteredRequest;
 import roomescape.reservation.entity.ReservationTime;
 import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.service.ReservationService;
