+++ src/test/java/roomescape/reservation/integration/ReservationTimeIntegrationTest.java
@@ -17,7 +17,7 @@ import roomescape.global.error.exception.ConflictException;
 import roomescape.member.entity.Member;
 import roomescape.member.entity.RoleType;
 import roomescape.member.repository.MemberRepository;
-import roomescape.reservation.dto.request.ReservationTimeRequest.ReservationTimeCreateRequest;
+import roomescape.reservation.dto.request.ReservationTimeCreateRequest;
 import roomescape.reservation.entity.Reservation;
 import roomescape.reservation.entity.ReservationTime;
 import roomescape.reservation.repository.ReservationRepository;
