+++ src/test/java/roomescape/theme/unit/service/ThemeServiceTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import roomescape.global.error.exception.ConflictException;
-import roomescape.theme.dto.request.ThemeRequest.ThemeCreateRequest;
+import roomescape.theme.dto.request.ThemeCreateRequest;
 import roomescape.theme.repository.ThemeRepository;
 import roomescape.theme.service.ThemeService;
